reafactor: moved reusable types to types directory, fixed small timeconverter bug, improved typesafety
