+++ src/app/api/ticker/route.ts
@@ -1,18 +1,16 @@
 import { API_ENDPOINTS } from "@/app/constants/urls";
+import { TickerData } from "@/types/TickerData";
 import axios from "axios";
 import { NextResponse } from "next/server";
 
-let cache: any = {
-  data: null,
-  timestamp: 0,
-};
+let cache: TickerData;
 
 const CACHE_DURATION = 30000; // 30 seconds
 
 export async function GET() { 
   const now = Date.now();
-  if (cache.data && now - cache.timestamp < CACHE_DURATION) {
-    return NextResponse.json(cache.data);
+  if (cache && now - cache.timestamp < CACHE_DURATION) {
+    return NextResponse.json(cache);
   }
 
   try {
@@ -37,21 +35,18 @@ export async function GET() {
         coinbase: coinbaseRate,
         bitfinex: bitfinexPrice,
       },
-      timestamp: now, // just to testing
-    };
-
-    cache = {
-      data: responseData,
       timestamp: now,
     };
+
+    cache = responseData;
 
     return NextResponse.json(responseData);
   } catch (error) {
     // console.error('Error fetching ticker data:', error);
 
-    if (cache.data) {
+    if (cache) {
       return NextResponse.json({
-        ...cache.data,
+        ...cache,
         error: 'Failed to fetch new data, returning cached data',
       });
     }
