+++ src/types/SelectedTradingPair.ts
@@ -1,0 +1,4 @@
+export interface SelectedTradingPair {
+  url_symbol: string, 
+  description: string
+}
