+++ src/types/TickerData.ts
@@ -1,0 +1,10 @@
+export interface TickerData {
+  averagePrice: string;
+  details: {
+    bitstamp: number;
+    coinbase: number;
+    bitfinex: number;
+  };
+  error?: string;
+  timestamp: number
+}
